Refactored sessions login
