+++ Dmitry_Selin/views/todolist.hbs
@@ -23,7 +23,7 @@
         <h4>Завершенные задачи</h4>
         <div id="div-tasks-done"></div>
       </div>
-      <script type="text/javascript" src="todolist.js"></script>
+      <script type="text/javascript" src="todo.js"></script>
       <script>
         window.onload = renderTasks({{{tasks}}});
       </script>
